requirements.txt
```text
streamlit==1.28.0
pandas==2.1.0
numpy==1.24.3
plotly==5.17.0
scikit-learn==1.3.0
Pillow==10.0.0
```
